+++ api/v1alpha1/database_types.go
@@ -41,8 +41,8 @@ type DatabaseStatus struct {
 // +kubebuilder:subresource:status
 
 // Database is the Schema for the databases API
-// +kubebuilder:printcolumn:name="Service Name",type="string",JSONPath=".spec.serviceName"
 // +kubebuilder:printcolumn:name="Project",type="string",JSONPath=".spec.project"
+// +kubebuilder:printcolumn:name="Service Name",type="string",JSONPath=".spec.serviceName"
 type Database struct {
 	metav1.TypeMeta   `json:",inline"`
 	metav1.ObjectMeta `json:"metadata,omitempty"`