+++ controllers/grafana_controller.go
@@ -26,9 +26,9 @@ func newGrafanaReconciler(c Controller) reconcilerType {
 	return &GrafanaReconciler{Controller: c}
 }
 
-// +kubebuilder:rbac:groups=aiven.io,resources=grafanas,verbs=get;list;watch;create;update;patch;delete
-// +kubebuilder:rbac:groups=aiven.io,resources=grafanas/status,verbs=get;update;patch
-// +kubebuilder:rbac:groups=aiven.io,resources=grafanas/finalizers,verbs=get;list;watch;create;update;patch;delete
+//+kubebuilder:rbac:groups=aiven.io,resources=grafanas,verbs=get;list;watch;create;update;patch;delete
+//+kubebuilder:rbac:groups=aiven.io,resources=grafanas/status,verbs=get;update;patch
+//+kubebuilder:rbac:groups=aiven.io,resources=grafanas/finalizers,verbs=get;create;update
 
 func (r *GrafanaReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
 	return r.reconcileInstance(ctx, req, newGenericServiceHandler(newGrafanaAdapter), &v1alpha1.Grafana{})