+++ controllers/project_controller.go
@@ -31,9 +31,9 @@ func newProjectReconciler(c Controller) reconcilerType {
 // ProjectHandler handles an Aiven project
 type ProjectHandler struct{}
 
-// +kubebuilder:rbac:groups=aiven.io,resources=projects,verbs=get;list;watch;create;update;patch;delete
-// +kubebuilder:rbac:groups=aiven.io,resources=projects/status,verbs=get;update;patch
-// +kubebuilder:rbac:groups=aiven.io,resources=projects/finalizers,verbs=get;list;watch;create;update;patch;delete
+//+kubebuilder:rbac:groups=aiven.io,resources=projects,verbs=get;list;watch;create;update;patch;delete
+//+kubebuilder:rbac:groups=aiven.io,resources=projects/status,verbs=get;update;patch
+//+kubebuilder:rbac:groups=aiven.io,resources=projects/finalizers,verbs=get;create;update
 
 func (r *ProjectReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
 	return r.reconcileInstance(ctx, req, ProjectHandler{}, &v1alpha1.Project{})