+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add `KafkaSchemaRegistryACL` kind
 - Add `ClickhouseDatabase` kind
 - Fix secret creation for kinds with no secrets
+- Include the Kubernetes version in the Go client's user agent
 - Replace `Database` kind validations and default values with CRD validation rules
 - Perform upgrade tasks to check if PG service can be upgraded before updating the service
 - Expose project CA certificate to service secrets: `REDIS_CA_CERT`, `MYSQL_CA_CERT`, etc.